% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npm.R
\name{npm_audit}
\alias{npm_audit}
\title{NPM Audit Fix}
\usage{
npm_audit(fix = FALSE)
}
\arguments{
\item{fix}{Whether to also fix issues.}
}
\value{
Invisibly returns the output of the command
as a \code{character} vector.
}
\description{
Audit NPM packages and fix potential issues.
}
\examples{
\dontrun{npm_audit()} 

}
