% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npm.R
\name{npm_init}
\alias{npm_init}
\title{NPM Init}
\usage{
npm_init()
}
\value{
Invisibly returns the output of the command
as a \code{character} vector.
}
\description{
Initialise an NPM project.
}
\examples{
\dontrun{npm_init()}

}
