% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npm.R
\name{npm_install}
\alias{npm_install}
\title{NPM Install}
\usage{
npm_install(..., scope = c("dev", "prod", "global"))
}
\arguments{
\item{...}{Names of packages to install.}

\item{scope}{Scope of the installation of the packages.}
}
\description{
NPM Install
}
