% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npm.R
\name{npm_outdated}
\alias{npm_outdated}
\title{Outdated Packages}
\usage{
npm_outdated()
}
\value{
The output of the command
as a \code{character} vector.
}
\description{
Get outdated NPM packages.
}
\examples{
\dontrun{npm_outdated()} 

}
