% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npm.R
\name{npm_path}
\alias{npm_path}
\alias{npm_path_set}
\alias{npm_path_get}
\title{Set Path to NPM Installation}
\usage{
npm_path_set(path = Sys.getenv("NPM_PATH"))

npm_path_get()
}
\arguments{
\item{path}{Path to NPM binary.}
}
\description{
Define the path to the NPM installation. By default,
this should not be necessary, parcel will find the path
by itself if NPM was installed correctly.
}
\section{Functions}{
\itemize{
\item \code{npm_path_get}: Get NPM path
}}

