% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npm.R
\name{npm_update}
\alias{npm_update}
\title{Update Packages}
\usage{
npm_update()
}
\value{
Invisibly returns the output of the command
as a \code{character} vector.
}
\description{
Update NPM packages.
}
\examples{
\dontrun{npm_update()} 

}
