% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npm.R
\name{npm_run}
\alias{npm_run}
\title{NPM Run}
\usage{
npm_run(...)
}
\arguments{
\item{...}{arguments to pass to the \code{npm} command.}
}
\value{
Invisibly returns the output of the command.

Invisibly returns the output of the command
as a \code{character} vector.
}
\description{
NPM Run
}
